Fixed bug where new author names weren't displaying.
After using testing, discovered that adding a new recipe by a user with
no other recipes does not update the list of authors that gets outputted
during search.
This was because of the issue that the list was from a variable in the
main file, so when I imported it, there were two versions - the original
and the one in all other files.
I moved the function that references it to another file so it references
the updated version.